requirements.txt
boto3
hvac
pyyaml
